requirements.txt
Flask
torch
pandas
Pillow
ultralytics
